requirements.txt
fastapi
uvicorn
transformers
psycopg2-binary
sqlalchemy
sqlalchemy-cockroachdb
asyncpg
